Return nested compartment OCID
For customers using nested compartments the ocid wasnt returned. As per the OCI CLI doc - With the exception of the tenancy (root compartment), the ListCompartments operation returns only the first-level child compartments in the parent compartment specified in compartmentId. The list does not include any subcompartments of the child compartments (grandchildren).